API is working

+++ back/src/main/java/com/openclassrooms/mddapi/repository/ArticleRepository.java
@@ -3,4 +3,4 @@ package com.openclassrooms.mddapi.repository;
 import com.openclassrooms.mddapi.entity.Article;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ArticleRepository extends CrudRepository<Article, Integer> {}
+public interface ArticleRepository extends CrudRepository<Article, Long> {}
