+++ back/src/main/java/com/openclassrooms/mddapi/repository/PostRepository.java
@@ -3,4 +3,4 @@ package com.openclassrooms.mddapi.repository;
 import com.openclassrooms.mddapi.entity.Post;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PostRepository extends CrudRepository<Post, Integer> {}
+public interface PostRepository extends CrudRepository<Post, Long> {}
