+++ back/src/main/java/com/openclassrooms/mddapi/repository/TopicRepository.java
@@ -3,4 +3,4 @@ package com.openclassrooms.mddapi.repository;
 import com.openclassrooms.mddapi.entity.Topic;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TopicRepository extends CrudRepository<Topic, Integer> {}
+public interface TopicRepository extends CrudRepository<Topic, Long> {}
