+++ back/src/main/java/com/openclassrooms/mddapi/repository/UserRepository.java
@@ -3,4 +3,4 @@ package com.openclassrooms.mddapi.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.openclassrooms.mddapi.entity.User;
 
-public interface UserRepository extends CrudRepository<User, Integer> {}
+public interface UserRepository extends CrudRepository<User, Long> {}
